requirements.txt
Flask
gunicorn
numpy
opencv-python
Pillow
pillow-heif
rembg
starlette


